loader: use local moon images instead of NASA's API
luna no longer requires an internet connection at runtime.  Instead, all
of NASA's hourly moon images from 2021 and 2022 are archived, along with
their moon ephemeris data (mooninfo.json).  50 years of future ephemeris
data from JPL Horizons is also archived.  To get the current moon image
to display, the date is looked up in the ephemeris data, getting the
moon's phase and libration parameters.  Then, the closest matching image
from the archive is used.

The 17520 saved images, covering two years of moon cycles, have a
good-enough match for every hour of the 50 years: if not identical, then
with similar-enough parameters as to be visually close.

This trades a lot of static data for not needing an internet connection,
which makes luna more stable and more flexible.